requirements.txt
weaviate-client
yt-dlp
python-dotenv